librkvpu: encoder: h264: merge H.264 encoder
I am not sure whether it would work, the old API is really buggy,
it even doesn't implement inheritance in right way.

And there are some APIs in rate control seems not been used anymore,
the implemation can't be found on the code anymore, all
the references to those funtions are been comment or condition
disable by pre-processor. It would be safe to remove those code.

Signed-off-by: Randy Li <randy.li@rock-chips.com>